// Ref/Web/ReferenceDataUpdateService.Web/src/models/RefApplicationAttributeTypeEnum.ts
export enum RefApplicationAttributeTypeEnum {
	Number = "Number",
	String = "String",
	Boolean = "Boolean",
	DateTime = "DateTime",
	File = "File",
	StatusFlag = "StatusFlag",
	Credential = "Credential",
	SecretFile = "SecretFile",
	Default = "Default"
}

export default RefApplicationAttributeTypeEnum;